requirements.txt
azure-functions
pandas
requests
openpyxl
chardet
beautifulsoup4
